+++ R/SLOPE.R
@@ -150,8 +150,16 @@
 #' @param max_passes maximum number of passes for optimizer
 #' @param diagnostics should diagnostics be saved for the model fit (timings,
 #'   primal and dual objectives, and infeasibility)
-#' @param screening whether the strong rule for SLOPE be used to screen
-#'   variables for inclusion
+#' @param screen whether to use predictor screening rules
+#' @param screen_set what type of screening algorithm to use.
+#'   * `"working"` is the approach used in glmnet, where
+#'     the previously active set is used as a working set that is
+#'     iteratively expanded by examining violations to the KKT rules in the
+#'     strong set and when there are no failures in the strong set, checked
+#'     against the full set of predictors; this approach is
+#'     particularly useful when there is considerable correlation
+#'   * `"strong"` uses the strong set as working set and checks for KKT
+#'     violations in the full set of predictors
 #' @param verbosity level of verbosity for displaying output from the
 #'   program. Setting this to 1 displays basic information on the path level,
 #'   2 a little bit more information on the path level, and 3 displays
@@ -281,7 +289,8 @@ SLOPE <- function(x,
                   lambda_min_ratio = if (n < p) 1e-2 else 1e-4,
                   n_sigma = 100,
                   q = 0.1*min(1, n/p),
-                  screening = TRUE,
+                  screen = TRUE,
+                  screen_alg = c("working", "strong"),
                   tol_dev_change = 1e-5,
                   tol_dev_ratio = 0.995,
                   tol_abs = 1e-5,
@@ -323,6 +332,7 @@ SLOPE <- function(x,
   ocall <- match.call()
 
   family <- match.arg(family)
+  screen_alg <- match.arg(screen_alg)
 
   if (is.character(scale)) {
     scale <- match.arg(scale)
@@ -464,7 +474,8 @@ SLOPE <- function(x,
                   center = center,
                   n_sigma = n_sigma,
                   n_targets = n_targets,
-                  screening = screening,
+                  screen = screen,
+                  screen_alg = screen_alg,
                   sigma = sigma,
                   sigma_type = sigma_type,
                   lambda = lambda,