+++ tests/testthat/test-screening-rules.R
@@ -4,8 +4,8 @@ test_that("screening rules return correct results for instances with known viola
   for (family in c("gaussian", "binomial", "poisson", "multinomial")) {
     d <- SLOPE:::randomProblem(100, 10, q = 0.1, response = family)
 
-    fit0 <- SLOPE(d$x, d$y, family = family, screening = FALSE)
-    fit1 <- SLOPE(d$x, d$y, family = family, screening = TRUE)
+    fit0 <- SLOPE(d$x, d$y, family = family, screen = FALSE)
+    fit1 <- SLOPE(d$x, d$y, family = family, screen = TRUE)
 
     expect_equivalent(coef(fit0), coef(fit1), 1e-4)
   }
@@ -16,7 +16,7 @@ test_that("basic screening rule works", {
 
   xy <- SLOPE:::randomProblem(100, 10, q = 0.1)
 
-  fit <- SLOPE(xy$x, xy$y, screening = TRUE)
+  fit <- SLOPE(xy$x, xy$y, screen = TRUE)
 
   expect_lt(length(fit$active_sets[[1]]), ncol(xy$x))
 })